move common herd logic to base class
(cherry picked from commit 8d22fc8e5b12a41a486ab7a742d3a5dd8f71f7ab)
